requirements.txt
telethon==1.32.1
pillow==10.2.0
requests==2.31.0
openai==1.12.0
pywin32==306
python-dotenv==1.0.1
cryptg==0.4.0
aiohttp==3.9.3 